fix(speakers): move findCompanyInfo to server-safe speakers-data lib
generateMetadata runs on the server and cannot call helpers exported from
the client speaker-ui module.
